fix: address terminal auth review regressions (#975)
Preserve the production dist-server entrypoint layout, tolerate malformed unrelated cookies during terminal auth, and separate ttyd startup failures from auth/config errors in terminal proxy responses.
